LayoutTests/fast/forms/implicit-submission-expected.txt

Tests various combinations of form elements and how implicit submission works with them.

Single text input should submit: PASS
Multiple text inputs should not submit: PASS
Multiple text inputs with submit should submit: PASS
Multiple text inputs with submit disabled should not submit: PASS
Multiple text inputs and multiple submits, first submit disabled should submit: PASS
Text input and text area, text input focused should submit: PASS
Text input and text area and a submit button, text input focused should submit: PASS
Text input and text area and a disabled submit button, text input focused should not submit: PASS
Text input and checkbox, text input focused should submit: PASS
Text input and radio, text input focused should submit: PASS
Text input and text area, textarea focused should not submit: PASS
Text input and checkbox, checkbox focused should not submit: PASS
Text input and radio, radio focused should not submit: PASS
Single radio should not submit: PASS
Single checkbox should not submit: PASS
Single checkbox with a submit button should submit: PASS
Single checkbox with a submit disabled should not submit: PASS
Single select should not submit: PASS
Text field and single select, text focused should submit: PASS
Text field and single select, select focused should not submit: PASS
Isindex should submit: PASS
